test(calendar): recategorize and clean up browser tests
